refactor: Extract UI state management into a dedicated UIStateManager for improved modularity and maintainability
